refactored template and now includes shipment ID
